+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.6] Apri 16, 2020
+* Add new ValueNotifier<bool> switchState to listen to the switch state changes. So when parent widge change the switchState to true, lite_switch will animated to "On".
+when parent widget changet switchState to false, lite_switch will animated to "Off"
+It can use switchState if you app need want to control the switch.
+
 ## [1.0.5] Apri 14, 2020
 * Fix the value change event bug.